+++ src/lib.rs
@@ -8,13 +8,15 @@ pub mod infrastructure;
 pub use application::{
     AddRandomNeuronCommand, AddRandomNeuronError, AddRandomNeuronHandler, AddRandomSynapseCommand,
     AddRandomSynapseError, AddRandomSynapseHandler, Command, CommandHandler,
-    MutateRandomSynapseWeightCommand, MutateRandomSynapseWeightError,
-    MutateRandomSynapseWeightHandler, Query, QueryHandler, QueryResult, RemoveRandomNeuronCommand,
-    RemoveRandomNeuronError, RemoveRandomNeuronHandler, RemoveRandomSynapseCommand,
-    RemoveRandomSynapseError, RemoveRandomSynapseHandler,
+    MutateNeuronActivationError, MutateRandomNeuronActivationCommand,
+    MutateRandomNeuronActivationHandler, MutateRandomSynapseWeightCommand,
+    MutateRandomSynapseWeightError, MutateRandomSynapseWeightHandler, Query, QueryHandler,
+    QueryResult, RemoveRandomNeuronCommand, RemoveRandomNeuronError, RemoveRandomNeuronHandler,
+    RemoveRandomSynapseCommand, RemoveRandomSynapseError, RemoveRandomSynapseHandler,
 };
 pub use domain::{
-    Activation, Event, Network as DomainNetwork, Neuron, RandomNeuronAdded, RandomNeuronRemoved,
-    RandomSynapseAdded, RandomSynapseRemoved, Synapse, SynapseWeightMutated,
+    Activation, Event, Network as DomainNetwork, Neuron, NeuronActivationMutated,
+    RandomNeuronAdded, RandomNeuronRemoved, RandomSynapseAdded, RandomSynapseRemoved, Synapse,
+    SynapseWeightMutated,
 };
 pub use infrastructure::{EventStore, FileEventStore};